requirements.txt
numpy==1.13.1
blinkt==0.1.0
google-assistant-library==0.0.3